feat: Add user_id validation to attachment detach method
The `detach` method in the `Attachment` struct now accepts an additional `user_id` parameter to validate if the user has permission to delete the attachment. If the `user_id` does not match the `attached.user_id`, an error is returned with the message "You do not have permission to delete this attachment". This change ensures that only the authorized user can delete attachments.